Feature/R4SU-47 redirect when login and logout added (#33)
