+++ packages/client/src/app/components/header/header.tsx
@@ -1,16 +1,15 @@
 import React, { FC, MouseEventHandler, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { Link, NavLink } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+import { Link, NavLink, useNavigate } from 'react-router-dom';
 import classNames from 'classnames';
 import Button from '@/app/components/common/button/button';
 import UserInfo from '@/app/components/userInfo/userInfo';
 import Logo from '@/assets/images/logo.svg';
 import Moon from '@/assets/images/moon.svg';
 import Sun from '@/assets/images/sun.svg';
-import { RoutePaths as Paths } from '@/app/router/router';
+import { RoutePaths as Paths, RoutePaths } from '@/app/router/router';
 import style from './header.module.scss';
 import AuthAPI from '@/app/api/AuthAPI';
-import { signOut } from '@/app/store/slices/userSlice';
 import TUser from '@/const/dataTypes/dataTypes';
 import changeColorMode from '@/app/helpers/changeColorMode';
 
@@ -22,14 +21,10 @@ type THeaderProps = {
 const Header: FC<THeaderProps> = () => {
     const [imageForChangeColorMode, setimageForChangeColorMode] = useState(Moon);
     const user = useSelector(state => (state as { user: unknown }).user) as TUser;
-    const dispatch = useDispatch();
+    const navigate = useNavigate();
 
     const logout: MouseEventHandler = () => {
-        AuthAPI.logout().then(() => dispatch(signOut()));
-
-        // AuthAPI.logout()
-        //  .then(() => alert('success logout'))
-        // .catch((err) => console.log(err));
+        AuthAPI.logout(() => navigate(RoutePaths.LANDING));
     };
 
     const calculateLinkClass = (isActive: boolean) => {
